refactor(viewer): modify useHeatmapDrawing hook to HeatmapViewer
